Final_upload/requirements.txt
# API Testing Framework Dependencies
requests>=2.31.0
pytest>=7.4.0
pytest-html>=3.2.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
# allure-pytest>=2.13.2  # Commented out - using specific version below
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
jsonschema>=4.19.0
tenacity>=8.2.0
faker>=19.3.0

# Reporting & Logging
loguru>=0.7.0
rich>=13.5.0
allure-pytest==2.15.3  # Allure reporting for pytest

# Performance Testing
locust>=2.16.0

# CI/CD & Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Security
requests-toolbelt>=1.0.0
python-jose>=3.3.0

